12_base_tool/05_analysis_tool/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
polygon-api-client>=1.12.0
plotly>=5.18.0
pytz>=2023.3
python-dotenv>=1.0.0
pandas-market-calendars>=4.3.0
requests>=2.31.0
pyarrow>=14.0.0
matplotlib>=3.8.0
